Item Inventory
Added simple inventory. UI still needs proper sprites added. Using the InteractableObject script and setting the correct fields,  items can either be set to pickups or as "converters", that take an item from your inventory and returns something else. All items must be on the "All Items" list on the Inventory component, and the fields on each object must be set to the exact string name of their corresponding game objects.